# apps/backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "VRK API",
        "contact": {}
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/standards": {
            "get": {
                "description": "Returns standard records visible inside the authenticated customer session scope.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "equipment"
                ],
                "summary": "List standards registry",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/standard.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/standard.StandardResponse"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/standard.Meta"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/standard.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/standard.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates one standard registry record inside the authenticated customer organization.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "equipment"
                ],
                "summary": "Create standard registry record",
                "parameters": [
                    {
                        "description": "Standard payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/standard.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/standard.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/standard.StandardResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/standard.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/standard.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/standard.Response"
                        }
                    }
                }
            }
        },
        "/standards/{id}": {
            "get": {
                "description": "Returns one standard record visible inside the authenticated customer session scope.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "equipment"
                ],
                "summary": "Get standard registry record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Standard ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/standard.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/standard.StandardResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/standard.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/standard.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/standard.Response"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates one standard record inside the authenticated customer organization.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "equipment"
                ],
                "summary": "Update standard registry record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Standard ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Standard patch payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/standard.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/standard.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/standard.StandardResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/standard.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/standard.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/standard.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/standard.Response"
                        }
                    }
                }
            }
        },
        "/standards/{id}/archive": {
            "post": {
                "description": "Archives one standard record inside the authenticated customer organization/session scope.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "equipment"
                ],
                "summary": "Archive standard registry record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Standard ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/standard.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/standard.StandardResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/standard.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/standard.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/standard.Response"
                        }
                    }
                }
            }
        },
        "/standards/{id}/journals": {
            "get": {
                "description": "Returns metrology journal entries for one standard visible inside the authenticated customer session scope.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "equipment"
                ],
                "summary": "List standard journal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Standard ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/standard.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/standard.JournalResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/standard.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/standard.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/standard.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Appends one metrology journal entry to the standard visible inside the authenticated customer organization.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "equipment"
                ],
                "summary": "Create standard journal entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Standard ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Journal payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/standard.CreateJournalRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/standard.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/standard.JournalResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/standard.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/standard.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/standard.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/standard.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "standard.CreateJournalRequest": {
            "type": "object",
            "properties": {
                "attachmentUrl": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "documentNumber": {
                    "type": "string"
                },
                "executorOrganization": {
                    "type": "string"
                },
                "operationDate": {
                    "type": "string"
                },
                "operationType": {
                    "type": "string"
                },
                "validUntil": {
                    "type": "string"
                }
            }
        },
        "standard.CreateRequest": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "documentUrl": {
                    "type": "string"
                },
                "identifier": {
                    "type": "string"
                },
                "metrologicalCharacteristics": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "ownerLabel": {
                    "type": "string"
                },
                "serialNumber": {
                    "type": "string"
                },
                "standardType": {
                    "type": "string"
                },
                "subdivisionId": {
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                }
            }
        },
        "standard.JournalResponse": {
            "type": "object",
            "properties": {
                "attachmentUrl": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "documentNumber": {
                    "type": "string"
                },
                "executorOrganization": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "operationDate": {
                    "type": "string"
                },
                "operationType": {
                    "type": "string"
                },
                "validUntil": {
                    "type": "string"
                }
            }
        },
        "standard.Meta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "standard.OwnershipScopeResponse": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "string"
                },
                "scopeId": {
                    "type": "string"
                },
                "scopeType": {
                    "type": "string"
                }
            }
        },
        "standard.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/standard.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "standard.StandardResponse": {
            "type": "object",
            "properties": {
                "archivedAt": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "documentUrl": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "identifier": {
                    "type": "string"
                },
                "journalCount": {
                    "type": "integer"
                },
                "latestJournal": {
                    "$ref": "#/definitions/standard.JournalResponse"
                },
                "linkedMeasuringInstruments": {
                    "type": "integer"
                },
                "metrologicalCharacteristics": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "nextDueDate": {
                    "type": "string"
                },
                "organizationId": {
                    "type": "string"
                },
                "ownershipScope": {
                    "$ref": "#/definitions/standard.OwnershipScopeResponse"
                },
                "serialNumber": {
                    "type": "string"
                },
                "standardType": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "standard.UpdateRequest": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "documentUrl": {
                    "type": "string"
                },
                "identifier": {
                    "type": "string"
                },
                "metrologicalCharacteristics": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "ownerLabel": {
                    "type": "string"
                },
                "serialNumber": {
                    "type": "string"
                },
                "standardType": {
                    "type": "string"
                },
                "subdivisionId": {
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                }
            }
        }
    }
}